test: split color_by tests into separate file
Extract the 11 color_by tests from test_embeddings2d.py into a
dedicated test_embeddings2d__color_by.py module.
